Making small error handling

+++ src/builtins/builtin_export.c
@@ -1,14 +1,14 @@
-/******************************************************************************/
+/* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
 /*   builtin_export.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nleite-s <nleite-s@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ekechedz <ekechedz@student.42.de>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 14:02:23 by codespace         #+#    #+#             */
-/*   Updated: 2024/11/07 11:29:08 by nleite-s         ###   ########.fr       */
+/*   Updated: 2024/11/10 11:49:02 by ekechedz         ###   ########.fr       */
 /*                                                                            */
-/******************************************************************************/
+/* ************************************************************************** */
 
 #include "../../include/minishell.h"
 #include "../../include/parser.h"
@@ -71,7 +71,7 @@ static int	check_exp_args(char *arg, t_scmd *scmd, t_tools *t)
 
 static int	print_export_list(t_tools *t)
 {
-	int		i;
+	size_t	i; // changed from int to size_t
 	char	**envcpy;
 	char	*full_env;
